requirements.txt
--index-url https://pypi.org/simple
--extra-index-url https://download.pytorch.org/whl/cpu
flask==3.0.2
pillow==10.4.0
numpy==1.24.3
transformers==4.36.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch
torchvision
torchaudio
protobuf==4.25.2
requests==2.31.0 